backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
asyncio==3.4.3
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
celery==5.3.4
celery[redis]==5.3.4
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
comm==0.2.2
config==0.5.1
constantly==23.10.4
crispy-bootstrap5==2025.4
cryptography==45.0.4
daphne==4.2.1
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.9.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.9
echo==0.11.0
et_xmlfile==2.0.0
executing==2.2.0
Flask==3.1.1
frozenlist==1.7.0
gevent==25.5.1
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
gspread==6.2.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.26
langchain-core==0.3.68
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langsmith==0.4.4
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
MouseInfo==0.1.3
msgpack==1.1.1
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.46.0
nest-asyncio==1.6.0
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.3.1
ollama==0.5.1
openai==1.93.0
openpyxl==3.1.5
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.50
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyJWT==2.9.0
PyMsgBox==1.0.9
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
pytweening==1.2.0
pytz==2025.1
pywhatkit==5.4
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
redis==6.2.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
selection==0.0.21
selenium==4.34.2
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sklearn-model==0.0.6
smmap==5.0.2
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.46.1
stripe==12.1.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
Twisted==25.5.0
txaio==25.6.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
undetected-chromedriver==3.5.5
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket==0.2.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.20.1
zope.event==5.1
zope.interface==7.2
zstandard==0.23.0
